Deleted the unnecessary classes (Seller, Identity...) and fixed student class
