--- src/main/java/za/ac/cput/domain/StudentGender.java
@@ -1,13 +1,0 @@
-/*  Gender.java
-    Gender POJO class
-    Author: AsandaNdhlela (230614345)
-    Date: 11 May 2025
-*/
-
-package za.ac.cput.domain;
-//import za.ac.cput.domain.demography.Gender;
-
-public class StudentGender {
-    protected int studentId;
-    protected int genderId;
-}
