--- src/main/java/za/ac/cput/domain/StudentIdentity.java
@@ -1,12 +1,0 @@
-/*  Gender.java
-    Gender POJO class
-    Author: AsandaNdhlela (230614345)
-    Date: 11 May 2025
-*/
-
-package za.ac.cput.domain;
-
-public class StudentIdentity {
-    protected int studentId;
-    protected int identityId;
-}
